+++ common/io/Cargo.toml
@@ -8,6 +8,10 @@ edition = "2021"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
+[lib]
+doctest = false
+test = false
+
 [dependencies]
 common-exception= {path = "../exception"}
 bytes = "1.1.0"
