+++ common/io/src/prelude.rs
@@ -18,6 +18,7 @@ pub use bytes::BytesMut;
 pub use crate::binary_de::BinaryDe;
 pub use crate::binary_read::BinaryRead;
 pub use crate::binary_ser::BinarySer;
+pub use crate::binary_write::put_uvarint;
 pub use crate::binary_write::BinaryWrite;
 pub use crate::binary_write::BinaryWriteBuf;
 pub use crate::buf_read::BufReadExt;
