+++ common/io/tests/it/utils.rs
@@ -12,14 +12,15 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-pub mod prelude;
+use common_io::prelude::*;
 
-mod binary_de;
-mod binary_read;
-mod binary_ser;
-mod binary_write;
-mod buf_read;
-mod marshal;
-mod stat_buffer;
-mod unmarshal;
-mod utils;
+#[test]
+fn convert_test() {
+    assert_eq!(convert_byte_size(1_f64), "1 B");
+    assert_eq!(convert_byte_size(1022_f64), "1.02 KB");
+    assert_eq!(convert_byte_size(1022_f64 * 10000000f64), "10.22 GB");
+
+    assert_eq!(convert_number_size(1_f64), "1");
+    assert_eq!(convert_number_size(1022_f64), "1.02 thousand");
+    assert_eq!(convert_number_size(10222_f64), "10.22 thousand");
+}
